refactor(codec): extract buildWireSizeFun into CodecEmitterWireSize.kt (step 7)
Move the wireSize(value) emit out of CodecEmitter into a new package-internal
top-level file. A fully self-contained stateless leaf (no batchCounter,
codeGenerator or logger, no calls into staying helpers); buildFileSpec calls it
unqualified across the package, same pattern as steps 4-6.

CodecEmitter.kt 1911 -> 1709 lines (down ~82% from the 9408-line original).
This is the last clean verbatim leaf — what remains (orchestration + framedBy +
batch + partial) is bound by the mutable batchCounter and the
appendDecodeField/appendEncodeField dispatchers, so further breakup is a design
refactor, not a move. Byte-identical codegen verified by CodecSnapshotTest
(jvmTest UP-TO-DATE = identical KSP output); processor + codec-test green.
